move function to its own file for reuse
